Refactoring work
Doing some refactoring before working on implementation details of the scoring plugin:
- Move all common types into types.go
- Move all common logic which being used by all plugins to common-plugin-helpers.go
- Set more specific naming convention. for example variable named 'Name', changed to FilterPluginName.
