+++ pkg/noderesourcetopology/filter_test.go
@@ -1,12 +1,9 @@
 /*
 Copyright 2021 The Kubernetes Authors.
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
     http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -29,7 +26,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/klog/v2"
 	framework "k8s.io/kubernetes/pkg/scheduler/framework/v1alpha1"
 )
 
@@ -40,39 +36,6 @@ const (
 	notExistingNICResourceName = "vendor/notexistingnic"
 	containerName              = "container1"
 )
-
-func makePodByResourceList(resources *v1.ResourceList) *v1.Pod {
-	return &v1.Pod{
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Resources: v1.ResourceRequirements{
-						Requests: *resources,
-						Limits:   *resources,
-					},
-				},
-			},
-		},
-	}
-}
-
-func makeResourceListFromZones(zones topologyv1alpha1.ZoneList) v1.ResourceList {
-	result := make(v1.ResourceList)
-	for _, zone := range zones {
-		for _, resInfo := range zone.Resources {
-			resQuantity, err := resource.ParseQuantity(resInfo.Allocatable.String())
-			if err != nil {
-				klog.Errorf("Failed to parse %s", resInfo.Allocatable.String())
-				continue
-			}
-			if quantity, ok := result[v1.ResourceName(resInfo.Name)]; ok {
-				resQuantity.Add(quantity)
-			}
-			result[v1.ResourceName(resInfo.Name)] = resQuantity
-		}
-	}
-	return result
-}
 
 func makePodByResourceListWithManyContainers(resources *v1.ResourceList, containerCount int) *v1.Pod {
 	var containers []v1.Container
@@ -310,14 +273,17 @@ func TestNodeResourceTopology(t *testing.T) {
 	for _, obj := range nodeTopologies {
 		fakeInformer.Informer().GetStore().Add(obj)
 	}
+	lister := fakeInformer.Lister()
 
 	tm := TopologyMatch{
 		policyHandlers: PolicyHandlerMap{
 			topologyv1alpha1.SingleNUMANodePodLevel:       SingleNUMAPodLevelHandler,
 			topologyv1alpha1.SingleNUMANodeContainerLevel: SingleNUMAContainerLevelHandler,
 		},
-		namespaces: []string{metav1.NamespaceDefault},
-		lister:     fakeInformer.Lister(),
+		data: commonPluginsData{
+			namespaces: []string{metav1.NamespaceDefault},
+			pLister:    &lister,
+		},
 	}
 
 	for _, tt := range tests {
